+++ shared/src/commonMain/kotlin/ai/lerna/multiplatform/service/dto/GlobalTrainingWeightsItem.kt
@@ -4,10 +4,10 @@ import org.jetbrains.kotlinx.multik.ndarray.data.D2Array
 
 class GlobalTrainingWeightsItem {
 	var weights // (prediction(lenra_job), weights(lerna_job))
-			: Map<String, D2Array<Double>>? = null
+			: Map<String, D2Array<Float>>? = null
 	var mlId // database lerna_job.ml_id
 			: Long? = null
 	var mlName // database lerna_ml.model
 			: String? = null
-	var accuracy: Double? = null
+	var accuracy: Float? = null
 }
